Feature/mys 59 timesheet manager (#17)
* MYS-59: added timesheet manager feature

* fixed eslint errors
