+++ src/sections/work/job/job-new-edit-address.tsx
@@ -284,6 +284,7 @@ export function JobNewEditAddress() {
               id: '',
               region: '',
               name: '',
+              logo_url: '',
               email: '',
               contact_number: '',
               unit_number: '',
@@ -302,6 +303,7 @@ export function JobNewEditAddress() {
               id: address.id || '',
               region: '',
               name: address.name || '',
+              logo_url: address.logo_url || '',
               email: '',
               contact_number: address.phoneNumber || '',
               unit_number: '',
