+++ src/sections/work/job/job-new-edit-status-date.tsx
@@ -6,7 +6,7 @@ import { useFormContext } from 'react-hook-form';
 import { useQuery } from '@tanstack/react-query';
 import isSameOrAfter from 'dayjs/plugin/isSameOrAfter';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
-import { useMemo, useState, useEffect, useCallback } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 
 import Box from '@mui/material/Box';
 
@@ -26,9 +26,46 @@ export function JobNewEditStatusDate() {
   const { watch, setValue, getValues } = useFormContext();
   const startTime = watch('start_date_time');
   const endTime = watch('end_date_time');
-  
-  // Wrap workers in useMemo to prevent dependency changes on every render
-  const workers = useMemo(() => watch('workers') || [], [watch]);
+
+  // Get workers data and force re-render
+  const [workers, setWorkers] = useState<any[]>([]);
+
+  // Update workers state when form values change
+  const watchedWorkers = watch('workers');
+  useEffect(() => {
+    const currentWorkers = getValues('workers') || [];
+    setWorkers(currentWorkers);
+  }, [getValues, watchedWorkers]);
+
+  // Watch timesheet manager for controlled value
+  const selectedTimesheetManager = watch('timesheet_manager_id');
+
+  // Clear timesheet manager when workers change and selection is no longer valid
+  useEffect(() => {
+    if (selectedTimesheetManager && selectedTimesheetManager !== '') {
+      const validWorkers = workers.filter(
+        (worker: any) =>
+          worker.id &&
+          worker.id !== '' &&
+          worker.first_name &&
+          worker.first_name.trim() !== '' &&
+          worker.last_name &&
+          worker.last_name.trim() !== ''
+      );
+
+      // Only clear if we have valid workers but the selection is not among them
+      // This prevents clearing during initialization when workers array might be empty
+      if (validWorkers.length > 0) {
+        const isSelectionValid = validWorkers.some(
+          (worker) => worker.id === selectedTimesheetManager
+        );
+
+        if (!isSelectionValid) {
+          setValue('timesheet_manager_id', '');
+        }
+      }
+    }
+  }, [workers, selectedTimesheetManager, setValue]);
 
   const [shiftHour, setShiftHour] = useState<number | string>('');
   const [hasManuallyChangedEndDate, setHasManuallyChangedEndDate] = useState(false);
@@ -48,10 +85,7 @@ export function JobNewEditStatusDate() {
   const [isProcessingDateChange, setIsProcessingDateChange] = useState(false);
 
   // Fetch time-off requests for conflict checking
-  const {
-    data: timeOffRequests = [],
-    isLoading: timeOffLoading,
-  } = useQuery({
+  const { data: timeOffRequests = [], isLoading: timeOffLoading } = useQuery({
     queryKey: ['time-off-conflicts-for-date-change', startTime, endTime],
     queryFn: async () => {
       if (!startTime || !endTime) return [];
@@ -68,26 +102,25 @@ export function JobNewEditStatusDate() {
   });
 
   // Fetch worker schedules for conflict checking
-  const {
-    data: workerSchedules = { scheduledWorkers: [] },
-    isLoading: scheduleLoading,
-  } = useQuery({
-    queryKey: ['worker-schedules-for-date-change', startTime, endTime],
-    queryFn: async () => {
-      if (!startTime || !endTime) return { scheduledWorkers: [] };
+  const { data: workerSchedules = { scheduledWorkers: [] }, isLoading: scheduleLoading } = useQuery(
+    {
+      queryKey: ['worker-schedules-for-date-change', startTime, endTime],
+      queryFn: async () => {
+        if (!startTime || !endTime) return { scheduledWorkers: [] };
 
-      const startISO = dayjs(startTime).toISOString();
-      const endISO = dayjs(endTime).toISOString();
-      const url = `${endpoints.work.job}/check-availability?start_time=${encodeURIComponent(startISO)}&end_time=${encodeURIComponent(endISO)}`;
+        const startISO = dayjs(startTime).toISOString();
+        const endISO = dayjs(endTime).toISOString();
+        const url = `${endpoints.work.job}/check-availability?start_time=${encodeURIComponent(startISO)}&end_time=${encodeURIComponent(endISO)}`;
 
-      const response = await fetcher(url);
-      return {
-        scheduledWorkers: response?.scheduledWorkers || [],
-        success: response?.success || false,
-      };
-    },
-    enabled: !!startTime && !!endTime,
-  });
+        const response = await fetcher(url);
+        return {
+          scheduledWorkers: response?.scheduledWorkers || [],
+          success: response?.success || false,
+        };
+      },
+      enabled: !!startTime && !!endTime,
+    }
+  );
 
   // Check for conflicts when dates change
   const checkDateChangeConflicts = useCallback(
@@ -230,12 +263,7 @@ export function JobNewEditStatusDate() {
         setIsProcessingDateChange(false);
       }
     },
-    [
-      checkDateChangeConflicts,
-      setValue,
-      timeOffLoading,
-      scheduleLoading,
-    ]
+    [checkDateChangeConflicts, setValue, timeOffLoading, scheduleLoading]
   );
 
   // Handle conflict dialog confirmation
@@ -437,8 +465,8 @@ export function JobNewEditStatusDate() {
       <Field.Text
         fullWidth
         name="po_number"
-        label="PO #"
-        placeholder="Enter Purchase Order Number"
+        label="PO # | NW #"
+        placeholder="Enter PO # | NW #"
         slotProps={{ inputLabel: { shrink: true } }}
       />
 
@@ -525,6 +553,108 @@ export function JobNewEditStatusDate() {
           '& .MuiOutlinedInput-notchedOutline': {
             border: 'none',
           },
+        }}
+      />
+
+      <Field.AutocompleteWithAvatar
+        key={`timesheet-manager-${workers.length}-${workers.map((w) => w.id).join('-')}`}
+        fullWidth
+        name="timesheet_manager_id"
+        label={(() => {
+          const validWorkers = workers.filter(
+            (worker: any) =>
+              worker.id &&
+              worker.id !== '' &&
+              worker.first_name &&
+              worker.first_name.trim() !== '' &&
+              worker.last_name &&
+              worker.last_name.trim() !== ''
+          );
+          return validWorkers.length === 0 ? 'Add worker first' : 'Timesheet Manager *';
+        })()}
+        placeholder={(() => {
+          const validWorkers = workers.filter(
+            (worker: any) =>
+              worker.id &&
+              worker.id !== '' &&
+              worker.first_name &&
+              worker.first_name.trim() !== '' &&
+              worker.last_name &&
+              worker.last_name.trim() !== ''
+          );
+          return validWorkers.length === 0 ? 'Add worker first' : 'Select timesheet manager';
+        })()}
+        options={(() => {
+          const validWorkers = workers.filter(
+            (worker: any) =>
+              worker.id &&
+              worker.id !== '' &&
+              worker.first_name &&
+              worker.first_name.trim() !== '' &&
+              worker.last_name &&
+              worker.last_name.trim() !== ''
+          );
+          return validWorkers.map((worker: any) => ({
+            value: worker.id,
+            label: `${worker.first_name} ${worker.last_name}`,
+            photo_url: worker.photo_url || '',
+            first_name: worker.first_name,
+            last_name: worker.last_name,
+          }));
+        })()}
+        disabled={(() => {
+          const validWorkers = workers.filter(
+            (worker: any) =>
+              worker.id &&
+              worker.id !== '' &&
+              worker.first_name &&
+              worker.first_name.trim() !== '' &&
+              worker.last_name &&
+              worker.last_name.trim() !== ''
+          );
+          return validWorkers.length === 0;
+        })()}
+        value={(() => {
+          const validWorkers = workers.filter(
+            (worker: any) =>
+              worker.id &&
+              worker.id !== '' &&
+              worker.first_name &&
+              worker.first_name.trim() !== '' &&
+              worker.last_name &&
+              worker.last_name.trim() !== ''
+          );
+          const currentValue = selectedTimesheetManager;
+
+          // Clear the field if no valid workers or no current value
+          if (!currentValue || currentValue === '' || validWorkers.length === 0) {
+            return null;
+          }
+
+          // Find the selected worker in current valid workers
+          const selectedWorker = validWorkers.find((w) => w.id === currentValue);
+
+          // If the previously selected worker is not in the current list, clear the selection
+          if (!selectedWorker) {
+            // Clear the form value to prevent stale selections
+            setValue('timesheet_manager_id', '');
+            return null;
+          }
+
+          return {
+            value: selectedWorker.id,
+            label: `${selectedWorker.first_name} ${selectedWorker.last_name}`,
+            photo_url: selectedWorker.photo_url || '',
+            first_name: selectedWorker.first_name,
+            last_name: selectedWorker.last_name,
+          };
+        })()}
+        onChange={(event: any, newValue: any) => {
+          if (newValue) {
+            setValue('timesheet_manager_id', newValue.value);
+          } else {
+            setValue('timesheet_manager_id', '');
+          }
         }}
       />
